+++ CHANGELOG.md
@@ -22,6 +22,8 @@ Other blocks are:
 
 ## Unreleased
 
+### Added
+- **LIB** - Added possibility to overwrite original X and Y axes in class `Spectrum` through method `overwrite_axis`.
 ### Fixed
 - **DEV** - Incorrect use of `typing.NoReturn` could confuse syntax analyzers (ex: Pylance).
 - **BLS** - Added correct error message if an input file for a curve is not found, instead of a useless Python traceback.